requirements.txt
crawl4ai>=0.5.0
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.5.0
beautifulsoup4>=4.12.0
openai>=1.0.0
anthropic>=0.5.0
requests>=2.31.0
tomli>=2.0.1
